fix: Improvements on ground item tracking (#118)
* fix: Improved dropped item detection

in onItemSpawned, the inventory still has a dropped clue in it.

A clue can only appear at the max spawn time if you don't have a clue of the same tier on you, so it must be the same clue yet to leave your inventory if it's appeared.

* feat: Added equals/hash methods for ClueInstance

This doesn't currently consider clueIDs, as those are usually not relevant to if a clue is likely to be the same as another or not.

* feat: Separated out easy-elite from beginner/masters in ClueGroundManager

This is intended to reduce the amount of checks overall done, as these clues don't need additional identifying to track what they are, as that's known through their itemID.

* feat: Created WorldPointToClueInstances to hold all ground clue tracking

This holds different lists and sets to help reduce the amount of looping or querying required for everything.

* fix: Fix issues with dev items not working in ClueInstance

* fix: Return null when stacks change size when menu is up

* fix: Ensure all tiles that appear after LOADING get wiped

* Remove isTrackedClueOrTornClue; Fixes Mark option

* fix: Torn clue tracking

* fix: Clear tile loaded tracking on game tick

* fix: Ensure tile order is respected for clues on reload

* fix: Refresh cluesByWorldPoint on add/remove

This does a full clear + compiling from the two different stores of beginner/master and easy-elite lists. A revert from the original code in terms of adding more work, but generally a good step to get everything working again as an initial step.

* fix: Stop overriding non-relevant wps in clueByWorldPoint on add/remove

* fix: Inventory overlay (#120)

* fix: Null state in WorldPointToClueInstance

* fix: Ensure non-null items for looping in WPTCI

* fix: Don't reverse ground clues overlay

* fix: Timers clueInstancesWithQuantityAtWp firstEntry null check

* fix: remove empty ground clues on WP if empty

---------

Co-authored-by: thelope <thelopeosrs@gmail.com>
Co-authored-by: TheLope <150454754+TheLope@users.noreply.github.com>